+++ CHANGES.rst
@@ -10,6 +10,7 @@ development
 - [mw] Allow Ikofax syntax from URLs like ?datasource=depatisnet&query=ikofax:EP666666%2Fpn
 - [ui] Refactor syntax chooser data model
 - [ui] Properly propagate Ikofax mode to Liveview mode
+- [ui] Improve "Share this search" dialog
 
 
 2018-09-28 0.152.0